requirements.txt
feedparser>=6.0
google-auth-oauthlib>=1.2
google-api-python-client>=2.130
python-telegram-bot>=21.0
google-generativeai>=0.8
anthropic>=0.40
openai>=1.50
pyyaml>=6.0
python-dotenv>=1.0
requests>=2.32
